# gateway-service/docs/swagger.yaml
basePath: /api
definitions:
...
  model.RequestAddPayment:
    properties:
      amount:
        type: number
      order_id:
        type: integer
      payment_date:
        type: string
      status:
        type: string
    type: object
  model.RequestAddProduct:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: number
      stok:
        type: integer
    type: object
...
  model.RequestUpdateProduct:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: number
      product_id:
        type: integer
      stok:
        type: integer
    type: object
...
  model.Response:
    properties:
      data: {}
      error_code:
        type: string
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  model.ResponseDetailProduct:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      string:
        type: integer
    type: object
...
  model.ResponseSuccessPayment:
    properties:
      amount:
        type: number
      payment_date:
        type: string
      payment_id:
        type: integer
      status:
        type: string
    type: object
  model.ResponseSucessProduct:
    properties:
      category:
        type: string
      description:
        type: string
      messege:
        type: string
      name:
        type: string
      price:
        type: number
      string:
        type: integer
    type: object
...
host: 43.218.113.72:8081
info:
  contact: {}
  description: Service API in Go Using Gin framework
  title: Synapsis E-commerce API
paths:
...
  /api/payment/add:
    post:
      description: Add order to payment
      parameters:
      - description: Payment Order
        in: body
        name: cary
        required: true
        schema:
          $ref: '#/definitions/model.RequestAddPayment'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccessPayment'
      summary: Create Paymenr
      tags:
      - payment-service
  /api/payment/delete:
    delete:
      description: Delete Payment Order
      parameters:
      - description: Payment ID
        in: query
        name: payment_id
        required: true
        type: string
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Delete Payment
      tags:
      - payment-service
...
  /product/add:
    post:
      description: Create Product
      parameters:
      - description: Create Product
        in: body
        name: CreateProduct
        required: true
        schema:
          $ref: '#/definitions/model.RequestAddProduct'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSucessProduct'
      summary: Create Product
      tags:
      - product-service
  /product/detail:
    get:
      description: Detail Product
      parameters:
      - description: Product ID
        in: query
        name: product_id
        required: true
        type: string
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseDetailProduct'
      summary: Detail Product
      tags:
      - product-service
...
  /product/listby-category:
    get:
      description: List Product by Category
      parameters:
      - description: Category ID
        in: query
        name: category_id
        required: true
        type: string
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseDetailProduct'
      summary: List Product by Category
      tags:
      - product-service
  /product/update:
    post:
      description: Update Product
      parameters:
      - description: Update Product
        in: body
        name: UpdateProduct
        required: true
        schema:
          $ref: '#/definitions/model.RequestUpdateProduct'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSucessProduct'
      summary: Update Product
      tags:
      - product-service
...
swagger: "2.0"